requirements.txt
flask
requests
python-whois
reportlab
gunicorn
python-nmap